Refactor StatsInfo import path and add StatsDetailed class to domain for improved structure
