+++ libs/domain/stats.py
